refactor: move file conversion to audioutils file.
